Create a new pass onnx-rewrite
